+++ packages/augur-simplified/src/modules/common/top-nav.tsx
@@ -1,7 +1,6 @@
-import React, { useEffect, useState, useMemo, useRef } from 'react';
+import React, { useEffect, useState, useRef } from 'react';
 import { useLocation } from 'react-router';
 import Styles from './top-nav.styles.less';
-import ButtonStyles from './buttons.styles.less';
 import { Link } from 'react-router-dom';
 import Logo from './logo';
 import classNames from 'classnames';
@@ -23,7 +22,7 @@ import {
 } from '@augurproject/augur-comps';
 const { GearIcon, ThreeLinesIcon } = Icons;
 const { ConnectAccount } = CompsConnectAccount;
-const { SecondaryButton, TinyButton } = ButtonComps;
+const { SecondaryButton } = ButtonComps;
 const { parsePath, makePath } = PathUtils;
 const {
   MARKET,
@@ -35,35 +34,12 @@ const {
 
 export const SettingsButton = () => {
   const {
-    settings: { slippage, showInvalidMarkets, showLiquidMarkets },
+    settings: { showInvalidMarkets, showLiquidMarkets },
     actions: { updateSettings },
   } = useAppStatusStore();
   const { account } = useUserStore();
   const [open, setOpened] = useState(false);
-  const [customVal, setCustomVal] = useState('');
   const settingsRef = useRef(null);
-  const isSelectedArray = useMemo(() => {
-    let output = [false, false, false, false];
-    switch (slippage) {
-      case '0.5': {
-        output[0] = true;
-        break;
-      }
-      case '1': {
-        output[1] = true;
-        break;
-      }
-      case '2': {
-        output[2] = true;
-        break;
-      }
-      default: {
-        output[3] = true;
-        break;
-      }
-    }
-    return output;
-  }, [slippage]);
 
   useEffect(() => {
     const handleWindowOnClick = (event) => {
@@ -84,12 +60,6 @@ export const SettingsButton = () => {
     };
   });
 
-  useEffect(() => {
-    if (customVal === '' && !['0.5', '1', '2'].includes(slippage)) {
-      setCustomVal(slippage);
-    }
-  }, [slippage, customVal]);
-
   return (
     <div className={Styles.SettingsMenuWrapper}>
       <SecondaryButton action={() => setOpened(!open)} icon={GearIcon} />
@@ -97,84 +67,6 @@ export const SettingsButton = () => {
         <ul className={Styles.SettingsMenu} ref={settingsRef}>
           <li>
             <h2>Settings</h2>
-          </li>
-          <li>
-            <label>
-              Slippage Tolerance
-              {generateTooltip(
-                'The maximum percentage the price can change and still have your transaction succeed.',
-                'slippageToleranceInfo'
-              )}
-            </label>
-            <ul>
-              <li>
-                <TinyButton
-                  text="0.5%"
-                  action={() => {
-                    updateSettings({ slippage: '0.5' }, account);
-                    setCustomVal('');
-                  }}
-                  selected={isSelectedArray[0]}
-                  className={ButtonStyles.TinyTransparentButton}
-                />
-              </li>
-              <li>
-                <TinyButton
-                  text="1%"
-                  action={() => {
-                    updateSettings({ slippage: '1' }, account);
-                    setCustomVal('');
-                  }}
-                  selected={isSelectedArray[1]}
-                  className={ButtonStyles.TinyTransparentButton}
-                />
-              </li>
-              <li>
-                <TinyButton
-                  text="2%"
-                  action={() => {
-                    updateSettings({ slippage: '2' }, account);
-                    setCustomVal('');
-                  }}
-                  selected={isSelectedArray[2]}
-                  className={ButtonStyles.TinyTransparentButton}
-                />
-              </li>
-              <li>
-                <div
-                  className={classNames({
-                    [Styles.Selected]: isSelectedArray[3],
-                  })}
-                >
-                  <input
-                    type="number"
-                    step="0.1"
-                    value={customVal}
-                    onChange={(v) => {
-                      setCustomVal(v.target.value);
-                    }}
-                    onBlur={() => {
-                      if (customVal !== slippage) {
-                        if (
-                          customVal === '' ||
-                          isNaN(Number(customVal)) ||
-                          Number(customVal) > 1000 ||
-                          Number(customVal) <= 0
-                        ) {
-                          setCustomVal(slippage);
-                        } else {
-                          updateSettings({ slippage: customVal }, account);
-                        }
-                      }
-                    }}
-                    placeholder="custom"
-                    max="1000"
-                    min="0.1"
-                  />
-                  <span>%</span>
-                </div>
-              </li>
-            </ul>
           </li>
           <li>
             <label>
