+++ packages/augur-simplified/src/modules/market/trading-form.tsx
@@ -24,6 +24,7 @@ import {
   ButtonComps,
 } from '@augurproject/augur-comps';
 import { useTrackedEvents } from '../../utils/tracker';
+import { Slippage } from '../common/slippage';
 const { doTrade, estimateEnterTrade, estimateExitTrade } = ContractCalls;
 const { CloseIcon } = Icons;
 const {
@@ -457,6 +458,7 @@ const TradingForm = ({
           }
           isBuy={orderType === BUY}
         />
+        <Slippage />
         <InfoNumbers infoNumbers={formatBreakdown(isBuy, breakdown, ammCash)} />
         {isLogged && !isApprovedTrade && (
           <ApprovalButton
